[JENKINS-66938] Allow save of 2+ priority definitions
The tables to divs fixes in release 4.0.0 of the plugin had the side
effect of preventing the second or later save of priority definitions
that contained more than one priority.

In my example, I defined a "High" priority, a "Low" priority, and a
"Default" priority.  When trying to save that definition, it initially
succeeds, then fails later attempts to save it.

It is unclear why the assignment of a class to the HTML select element
would cause the Stapler response to include unexpected data.