Add UpstreamCauseStrategy test
Test only attempts to exercise getUpstreamCause and is unable to exercise
it because I can't see why the downstream job is not in the queue in
the test environment, when the same job definition in a running Jenkins
environment leaves it in the queue.

When I define the jobs in Jenkins interactively, I can use the Jenkins
script console to run println(Jenkins.get().getItems()) and I can see
the downstream job is in the queue.  However, not so for the test.

Maybe the ideas in the test are helpful for you as you try to make
further progress.

It seems like UpstreamTest is a good example of a test that invokes the
getUpstreamCause code already.  It may be a good place to explore further.