Make test results clearer by checking decision logger results
Since the API provides a decision logger and an adjustment was made to
the decision logger in the test setup, let's use that decision logger
to make the test expectations clearer and the assertions more precise.

Initialize the JenkinsRule only once and create the project only once,
since neither are modified substantially for the test.

Use a more precise name for the methods that are called before test
execution so that later readers understand the purpose of the methods
without reading their contents.

Use the common name 'j' for the JenkinsRule and a more descriptive name
for the FreeStyleProject variable name.