Check results from decision logger
Do not use wildcard import of org.junit.Assert because that includes the
deprecated assertThat method.  Use the assertThat method from hamcrest
because it is not deprecated and works very well with hamcrest assertions.

Name the strategy uniquely for each test by including the test name in
the strategy name.  The TestName Rule is a very handy way to assure that
test data is unique for each test method without needing to worry about
the uniqueness directly.

Split the Before methods so that they can be named more specifically.