Add test for PropertyBasedJobInclusionStrategy class (#420)
* Add test for PropertyBasedJobInclusionStrategy class

* Made cloudbeesFolder package-private

* Add constructor to JobGroup class

Added a constructor to initialize description, priority, and jobGroupStrategy fields.
This change allows creating JobGroup instances with specific values for these fields.

* Changed FreestyleProject from project to j

* Check results from decision logger

Do not use wildcard import of org.junit.Assert because that includes the
deprecated assertThat method.  Use the assertThat method from hamcrest
because it is not deprecated and works very well with hamcrest assertions.

Name the strategy uniquely for each test by including the test name in
the strategy name.  The TestName Rule is a very handy way to assure that
test data is unique for each test method without needing to worry about
the uniqueness directly.

Split the Before methods so that they can be named more specifically.

* Replace JobGroup changes with more test code

Don't need a new constructor when setters are already available for all
the attributes used in the new constructor.

* Rename getDescriptor and make it package protected for tests

Reduce the amount of change in the production object by not creating a
new method that overrides the method from the superclass.

Limit the accessibility of the new method so that it is package protected
and not available to classes outside its own package.

Remove two duplicated test methods

* Improve documentation and add tests for PropertyBasedJobInclusionStrategy

- Added detailed Javadoc comments for `getThisDescriptor` and `getDescriptor` methods.
- Enhanced inline comments to clarify the use of `@SuppressWarnings("unchecked")`.

* Add package protected comment

---------

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>