Add tests for PriorityJobProperty class (#437)
* Add test for PriorityJobProperty class

* introduced a new conditional to check if form is empty or null

* Remove main object change and adjust test to match

Move the creation of the mock object inside the single test class that
needs it.  No benefit to initializing it sooner than needed or keeping
it visible longer than needed.

* Extract jobGroup creation to a separate method

Avoid duplicating code that is common between uses.

Use a job group name that is distinct for each test so that the single
JenkinsRule instance does not risk having the tests collide in their
use of the job group.

* Move project creation into the two tests that need it

No need to create the project at the top level and then create another
project inside one of the two tests that use a project.  Create a project
in the tests that need one.

---------

Co-authored-by: Mark Waite <mark.earl.waite@gmail.com>